+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #  Tree Tracker
 
 ## Next
+- Added "Upload History" screen (from now on we will move trees added by this device to the History tab)
 - Replaced "Tree list" screen with "Entities" screen.
 - Replaced Secrets.xcconfig with Pouch config.